Extract Feature Popup to Ember Component
This relies on using Ember to render the popup instead of JQuery.
