+++ app/components/openlayers-map.js
@@ -1,5 +1,4 @@
 import Component from '@ember/component';
-import Overlay from 'ol/Overlay';
 import Map from 'ol/Map';
 import View from 'ol/View';
 import { transform } from 'ol/proj';
@@ -38,49 +37,9 @@ export default Component.extend({
       })
     });
 
-    // Grab the "popup" element and link it to an OpenLayers Overlay
-    // TODO: Extract to Component
-    let popup = new Overlay({
-      element: this.$("#map-popup")[0],
-      positioning: 'bottom-center',
-      stopEvent: false,
-      offset: [0, -10]
-    });
-    map.addOverlay(popup);
-
-    // Set up click handler on the map to activate the "onClick" handler
-    // for the feature closest to the click.
-    map.on('click', (event) => {
-      let features = map.getFeaturesAtPixel(event.pixel);
-
-      // TODO: If multiple features are near the click event, a popup
-      // should be used to select one
-      if (features.length > 1) {
-        console.warn("Not yet supported");
-      } else if (features.length == 1) {
-        let feature = features[0];
-        let coordinates = feature.getGeometry().getCoordinates();
-
-        
-        let $popup = this.$(popup.getElement());
-        $popup.find(".title").html(feature.get('title'));
-
-        // Set the popup center coordinates *after* updating the popup
-        // contents, otherwise the popup won't be centered on the correct
-        // position.
-        popup.setPosition(coordinates);
-
-        // Activate the onClick callback for the feature component
-        let onClick = feature.get('onClick');
-        if (onClick) {
-          onClick();
-        }
-      }
-    });
-
     this.set('map', map);
     // Pass an event to any sub-components waiting for the Map element
     // to be ready.
-    this.trigger('mapReady', map);
+    this.trigger('ready', map);
   }
 });
