+++ app/components/openlayers-map/wmts-tile-layer.js
@@ -44,7 +44,7 @@ export default Component.extend({
       matrixIds[z] = z;
     }
 
-    this.get('parentComponent').on('mapReady', (map) => {
+    this.get('parentComponent').on('ready', (map) => {
       let tileLayer = new TileLayer({
         // extent: this.get('extent'),
         opacity: this.get('opacity'),
