+++ app/templates/components/openlayers-map.hbs
@@ -2,6 +2,7 @@
 	tile-layer=(component "openlayers-map/tile-layer" parentComponent=this)
 	wmts-tile-layer=(component "openlayers-map/wmts-tile-layer" parentComponent=this)
 	vector-layer=(component "openlayers-map/vector-layer" parentComponent=this)
+	popup=(component "openlayers-map/popup" parentComponent=this)
 ) as |components|}}
 	{{yield components}}
 {{/with}}
