+++ app/templates/components/openlayers-map/popup.hbs
@@ -1,0 +1,2 @@
+<div class="popup-arrow"></div>
+<h4 class="title">{{title}}</h4>
